Beginning

+++ src/analyzer/analyzer.py
@@ -11,6 +11,8 @@ this_path = os.path.dirname(os.path.realpath(__file__))
 
 exclude_fn = r'^llvm\|^load\|^_tr_\|^_\.\|^OPENSSL_cleanse'
 
+analyzer = "target/release/analyzer"
+
 def setup_parser(parser):
   parser.add_argument('--bc', type=str, default="", help='The .bc file to analyze')
   parser.add_argument('--slice-depth', type=int, default=1, help='Slice depth')
@@ -39,39 +41,53 @@ def generate_temp_folder_name():
 
 def main(args):
   db = args.db
-  packages_functions = {}
-  temp_folder_name = generate_temp_folder_name()
-  os.mkdir(db.analysis_dir() + "/" + temp_folder_name)
 
-  # Run analyzer individually on each bc package
-  for bc_file in db.bc_files():
-    if args.bc == '' or args.bc in bc_file:
+  # Check the bc_files to run
+  bc_files_to_run = [bc_file for bc_file in db.bc_files() if args.bc == '' or args.bc in bc_file]
+
+  if len(bc_files_to_run) > 1:
+
+    # Generate temporary folder that is shared among all runs
+    temp_folder_name = generate_temp_folder_name()
+    os.mkdir(db.analysis_dir() + "/" + temp_folder_name)
+    packages_functions = {}
+
+    # Run analyzer individually on each bc package
+    for bc_file in bc_files_to_run:
       try:
-        functions = run_analyzer(db, bc_file, temp_folder_name, args)
+        functions = run_analyzer_on_one_of_bc_files(db, bc_file, temp_folder_name, args)
         packages_functions[bc_file] = functions
       except Exception as e:
         print(e)
         exit()
 
-  # Run feature extractor on all data generated
-  input_file = generate_feature_extract_input(db, packages_functions, temp_folder_name)
-  run_feature_extractor(db, input_file, args)
+    # Run feature extractor on all data generated
+    input_file = generate_feature_extract_input(db, packages_functions, temp_folder_name)
+    run_feature_extractor(db, input_file, args)
 
-  # Remove the temp folder
-  shutil.rmtree(db.analysis_dir() + "/" + temp_folder_name)
+    # Remove the temp folder
+    shutil.rmtree(db.analysis_dir() + "/" + temp_folder_name)
+
+  else:
+    bc_file = bc_files_to_run[0]
+    run_analyzer_on_the_only_bc_file(db, bc_file, args)
 
 
-def get_analyzer_args(db, bc_file, temp_folder, args):
+def get_analyzer_args(db, bc_file, args, temp_folder=None, extract_features=False):
   bc_name = ntpath.basename(bc_file)
 
   base_args = [
     bc_file,
     db.analysis_dir(),
     '--subfolder', bc_name,
-    '--slice-depth', str(args.slice_depth),
-    '--no-feature',
-    '--target-num-slices-map-file', f'{temp_folder}/{bc_name}.json'
+    '--slice-depth', str(args.slice_depth)
   ]
+
+  if extract_features == False:
+    base_args += ['--no-feature']
+
+  if temp_folder != None:
+    base_args += ['--target-num-slices-map-file', f'{temp_folder}/{bc_name}.json']
 
   if args.include_fn:
     base_args += ['--target-inclusion-filter', args.include_fn]
@@ -115,18 +131,23 @@ def get_analyzer_args(db, bc_file, temp_folder, args):
   return base_args
 
 
-def run_analyzer(db, bc_file, temp_folder, args):
-  print(f"Running analyzer on {os.path.basename(bc_file)}")
+def run_analyzer_on_one_of_bc_files(db, bc_file, temp_folder, args):
+  print(f"=== Running analyzer on {os.path.basename(bc_file)}... ===")
   bc_name = ntpath.basename(bc_file)
-
-  analyzer = "target/release/analyzer"
-  analyzer_args = get_analyzer_args(db, bc_file, temp_folder, args)
+  analyzer_args = get_analyzer_args(db, bc_file, args, temp_folder=temp_folder)
   cmd = [analyzer] + analyzer_args
   run = subprocess.run(cmd, cwd=this_path)
 
   output_file = db.analysis_dir() + "/" + temp_folder + "/" + bc_name + ".json"
   with open(output_file) as f:
     return json.load(f)
+
+
+def run_analyzer_on_the_only_bc_file(db, bc_file, args):
+  bc_name = ntpath.basename(bc_file)
+  analyzer_args = get_analyzer_args(db, bc_file, args, extract_features=True)
+  cmd = [analyzer] + analyzer_args
+  run = subprocess.run(cmd, cwd=this_path)
 
 
 def generate_feature_extract_input(db, packages_functions, temp_folder):
@@ -162,6 +183,7 @@ def get_extractor_args(db, input_file, args):
 
 
 def run_feature_extractor(db, input_file, args):
+  print(f"=== Running feature extraction... ===")
   extractor = "target/release/feature-extract"
   extractor_args = get_extractor_args(db, input_file, args)
   cmd = [extractor] + extractor_args
